refactor sniff to be in kering instead of Parser class method. This is to support new Streamer class for SPAC wrapping that needs to use sniff without circular import.
